refactor(feel): move msgpack converter from expression-language
The FeelToMessagePackTransformer needs to also be used by the dmn
module.
